Use local theads for dask array compute
Add a compute_args configuration parameter to the HedwigImages class
to allow configuration of the dask schedualer used for dask array
computation and ZARR operations.

This addresses a deadlock issue when the Dask array is using the
global schedualer that Prefect uses. This avoids workers recusivedly
using the globlal schedualer while blocking a worker.